Support updating firmware from picorom tool
