+++ host/picorom/src/main.rs
@@ -2,14 +2,24 @@ use anyhow::{anyhow, Result};
 use clap::{Parser, Subcommand};
 use indicatif::{ProgressBar, ProgressStyle};
 use std::fs;
+use std::io::{self, Write};
 use std::iter;
 use std::path::{Path, PathBuf};
+use std::thread::sleep;
 use std::time::Duration;
 
-use picolink::*;
+use picolink::{
+    enumerate_picos, find_pico, get_device_location, wait_for_bootloader_at_location,
+    PicobootConnection, FLASH_SECTOR_SIZE,
+};
 
+mod firmware;
 mod rom_size;
+mod uf2;
+
+use crate::firmware::{upload_firmware, ProgressKind};
 use crate::rom_size::*;
+use crate::uf2::Uf2File;
 
 fn read_file(name: &Path, rom_size: RomSize) -> Result<Vec<u8>> {
     let mut data = fs::read(name)?;
@@ -124,6 +134,20 @@ enum Commands {
 
     /// Reboot the device into USB mode
     USBBoot { name: String },
+
+    /// Upload firmware to a PicoROM device
+    Firmware {
+        /// PicoROM device name
+        name: String,
+        /// Path to UF2 firmware file
+        firmware: PathBuf,
+        /// Skip confirmation prompt
+        #[arg(short = 'y', long)]
+        yes: bool,
+        /// Don't reboot after flashing
+        #[arg(long)]
+        no_reboot: bool,
+    },
 }
 
 fn main() -> Result<()> {
@@ -256,6 +280,166 @@ fn main() -> Result<()> {
             let mut pico = find_pico(&name)?;
             println!("Requesting USB boot");
             pico.usb_boot()?;
+        }
+
+        Commands::Firmware {
+            name,
+            firmware,
+            yes,
+            no_reboot,
+        } => {
+            // Parse and validate UF2 file
+            let uf2 = Uf2File::parse(&firmware)?;
+
+            let (start_addr, end_addr) = uf2
+                .address_range()
+                .ok_or_else(|| anyhow!("UF2 file contains no data"))?;
+
+            // Show summary
+            println!("Firmware: {:?}", firmware);
+            println!(
+                "  Blocks: {}, Total size: {} bytes",
+                uf2.block_count,
+                uf2.total_bytes()
+            );
+            println!("  Address range: 0x{:08X} - 0x{:08X}", start_addr, end_addr);
+
+            let sectors = uf2.sectors_to_erase(FLASH_SECTOR_SIZE);
+            let total_erase: u32 = sectors.iter().map(|(_, s)| s).sum();
+            println!(
+                "  Sectors to erase: {} ({} bytes)",
+                sectors.len(),
+                total_erase
+            );
+
+            // Confirmation prompt
+            if !yes {
+                print!("\nFlash firmware to '{}'? [y/N] ", name);
+                io::stdout().flush()?;
+
+                let mut input = String::new();
+                io::stdin().read_line(&mut input)?;
+
+                if !input.trim().eq_ignore_ascii_case("y") {
+                    println!("Aborted.");
+                    return Ok(());
+                }
+            }
+
+            // Try to find device and send to bootloader, or connect to existing bootloader
+            let mut conn = match find_pico(&name) {
+                Ok(mut pico) => {
+                    // Capture USB port location before rebooting
+                    // This allows us to identify the same physical device in bootloader mode
+                    let (bus_id, port_chain) = get_device_location(&name)?;
+
+                    println!("\nSending '{}' to bootloader...", name);
+                    pico.usb_boot()?;
+
+                    // Wait for device to disconnect
+                    sleep(Duration::from_millis(500));
+
+                    // Wait for bootloader to appear at the same USB port location
+                    let spinner = ProgressBar::new_spinner()
+                        .with_prefix("Waiting for bootloader")
+                        .with_style(
+                            ProgressStyle::with_template("{prefix:.bold} {spinner} {msg}")
+                                .unwrap()
+                                .tick_chars(r"\|/--"),
+                        );
+                    spinner.enable_steady_tick(Duration::from_millis(100));
+
+                    let conn = wait_for_bootloader_at_location(
+                        &bus_id,
+                        &port_chain,
+                        Duration::from_secs(10),
+                    )?;
+                    spinner.finish_with_message("Connected");
+                    conn
+                }
+                Err(_) => {
+                    // Device not found - check if there's already a bootloader device
+                    println!("\nDevice '{}' not found, checking for bootloader...", name);
+                    match PicobootConnection::open(None) {
+                        Ok(conn) => {
+                            println!("Found bootloader device: {}", conn.device_id);
+                            conn
+                        }
+                        Err(_) => {
+                            return Err(anyhow!(
+                                "Device '{}' not found and no bootloader device available",
+                                name
+                            ));
+                        }
+                    }
+                }
+            };
+
+            // Erase progress bar
+            let erase_progress = ProgressBar::new(total_erase as u64)
+                .with_prefix("Erasing flash")
+                .with_style(
+                    ProgressStyle::with_template(
+                        "{prefix:.bold} [{wide_bar:.yellow/red}] {bytes}/{total_bytes}",
+                    )
+                    .unwrap()
+                    .progress_chars("#>-"),
+                );
+
+            // Write progress bar
+            let write_progress = ProgressBar::new(uf2.total_bytes() as u64)
+                .with_prefix("Writing flash")
+                .with_style(
+                    ProgressStyle::with_template(
+                        "{prefix:.bold} [{wide_bar:.cyan/blue}] {bytes}/{total_bytes}",
+                    )
+                    .unwrap()
+                    .progress_chars("#>-"),
+                );
+
+            // Upload firmware
+            upload_firmware(&uf2, &mut conn, |kind, current, _total| match kind {
+                ProgressKind::Erase => erase_progress.set_position(current),
+                ProgressKind::Write => write_progress.set_position(current),
+            })?;
+
+            erase_progress.finish();
+            write_progress.finish();
+
+            if !no_reboot {
+                println!("Rebooting device...");
+                conn.reboot(500)?;
+
+                // Wait for device to come back
+                sleep(Duration::from_millis(1000));
+
+                let spinner = ProgressBar::new_spinner()
+                    .with_prefix("Waiting for device")
+                    .with_style(
+                        ProgressStyle::with_template("{prefix:.bold} {spinner} {msg}")
+                            .unwrap()
+                            .tick_chars(r"\|/--"),
+                    );
+                spinner.enable_steady_tick(Duration::from_millis(100));
+
+                // Try to reconnect
+                let deadline = std::time::Instant::now() + Duration::from_secs(10);
+                loop {
+                    if let Ok(_pico) = find_pico(&name) {
+                        spinner.finish_with_message("Device online");
+                        break;
+                    }
+                    if std::time::Instant::now() >= deadline {
+                        spinner.finish_with_message("Timeout (device may still boot)");
+                        break;
+                    }
+                    sleep(Duration::from_millis(200));
+                }
+
+                println!("\nFirmware update complete!");
+            } else {
+                println!("\nFirmware written. Device left in bootloader mode.");
+            }
         }
     }
